Azure/requirements.txt
flask==2.3.2
tensorflow==2.18.0  
numpy==1.26.4
pillow==11.0.0
gunicorn==20.1.0
fastapi==0.112.2
python-multipart==0.0.20
uvicorn==0.22.0